third_party/blink/web_tests/virtual/html-anchor-attribute-disabled/external/wpt/html/semantics/popovers/popover-stacking-anchor-attribute.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] anchor attribute relationship
  assert_true: expected true got false
[FAIL] indirect anchor attribute relationship
  assert_true: expected true got false
[FAIL] more complex nesting, all using anchor ancestry
  assert_equals: Popover #1 incorrect state expected true but got false
Harness: the test ran to completion.
